GCB-179: refactor EventsSource and CalendarNameSource
