+++ GrandCentralBoard/Widgets/Watch/Server/Calendar.swift
@@ -14,4 +14,11 @@ struct Calendar : Decodable {
     static func decode(json: AnyObject) throws -> Calendar {
         return try Calendar(name: json => "summary")
     }
-}
+
+    static func decodeFromData(data: NSData) throws -> Calendar {
+        if let jsonResult = try NSJSONSerialization.JSONObjectWithData(data, options: []) as? NSDictionary {
+            return try Calendar.decode(jsonResult)
+        }
+
+        throw EventsError.WrongFormat
+    }}
