+++ GrandCentralBoard/Widgets/Watch/Source/EventsSource.swift
@@ -10,11 +10,24 @@ import GrandCentralBoardCore
 
 class EventsSource : Asynchronous {
     typealias ResultType = Result<[Event]>
-
-    let interval: NSTimeInterval = 60
     let sourceType: SourceType = .Momentary
 
+    let calendarID: String
+    let interval: NSTimeInterval
+    let dataProvider: CalendarDataProviding
+
+    init(calendarID: String, dataProvider: CalendarDataProviding, refreshInterval: NSTimeInterval = 60) {
+        self.calendarID = calendarID
+        self.interval = refreshInterval
+        self.dataProvider = dataProvider
+    }
+
     func read(closure: (ResultType) -> Void) {
-        fatalError("Not implemented")
+        dataProvider.fetchEventsForCalendar(calendarID) { result in
+            switch result {
+            case .Success(let events): closure(.Success(events))
+            case .Failure(let error): closure(.Failure(error))
+            }
+        }
     }
 }
